--- a/optimum/habana/transformers/generation/utils.py
+++ b/optimum/habana/transformers/generation/utils.py
@@ -2301,7 +2301,7 @@ class GaudiGenerationMixin(GenerationMixin):
             ):
                 # Pad the returned pask key values tensors from prefill phase forward run to maximum length
                 # before starting the decode phase.
-                if outputs.past_key_values[0][0].shape[2] ==  model_inputs['input_ids']:
+                if outputs.past_key_values[0][0].shape[2] ==  model_inputs['input_ids'].shape[1]:
                     self._pad_past_key_values(model_kwargs)
                 model_kwargs["pad_done"] = True